Fixed time in pressure
